pyproject.toml
[tool.poetry]
name = "crm-epicevents"
version = "0.1.0"
description = "CRM pour Epic Events"
authors = ["Julien Denizot <luunz71@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.36"
pymysql = "^1.1.1"
python-decouple = "^3.8"
passlib = "^1.7.4"
bcrypt = "^4.2.1"
alembic = "^1.14.0"
pytest = "^8.3.4"
argon2-cffi = "^23.1.0"
pyjwt = "^2.10.1"
argon2 = "^0.1.10"
click = "^8.1.7"
rich = "^13.9.4"
prompt-toolkit = "^3.0.48"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
